Renaming some files for consistency with naming conventions.
